+++ palladian-retrieval/src/test/java/ws/palladian/retrieval/search/web/GoogleCustomSearcherTest.java
@@ -23,7 +23,7 @@ public class GoogleCustomSearcherTest {
         assertEquals("http://en.wikipedia.org/wiki/Palladian_architecture", results.get(0).getUrl());
         assertEquals("Palladian architecture is a European style of architecture derived from the   designs of the Venetian architect Andrea Palladio (1508–1580). The term \"  Palladian\" ...", results.get(0).getSummary());
         
-        int resultCount = GoogleCustomSearcher.parseResultCount(jsonString);
+        long resultCount = GoogleCustomSearcher.parseResultCount(jsonString);
         assertEquals(147000, resultCount);
     }
